security: Fix vulnerability by switching websocket library
Currently this crate uses the `websocket` crate, that has a dependency
on a hyper version that is subject to a security vulnerability.
This change switches the implementation to the `tokio-tungstenite`
websocket library, which exposes an async interface. In order to fit
the async interface with the current sync interface, async-bridging is
used